refactor useCohortHandler
