+++ src/common/handlers/cohorts.js
@@ -195,3 +195,118 @@ export const generateCohortSyllabusModules = async (id) => {
     return {};
   }
 };
+
+export const nestAssignments = ({
+  id, label = '', read, practice, project, answer, taskTodo = [],
+}) => {
+  const getTaskProps = (slug) => taskTodo.find(
+    (task) => task.associated_slug === slug,
+  );
+  const currentDate = new Date();
+
+  const updatedRead = read?.map((el) => ({
+    ...el,
+    id,
+    label,
+    task_status: getTaskProps(el.slug)?.task_status || '',
+    revision_status: getTaskProps(el.slug)?.revision_status || '',
+    created_at: getTaskProps(el.slug)?.created_at || '',
+    position: el.position,
+    type: 'Read',
+    icon: 'book',
+    task_type: 'LESSON',
+  })).sort((a, b) => b.position - a.position);
+
+  const updatedPractice = practice?.map((el) => ({
+    ...el,
+    id,
+    label,
+    task_status: getTaskProps(el.slug)?.task_status || '',
+    revision_status: getTaskProps(el.slug)?.revision_status || '',
+    created_at: getTaskProps(el.slug)?.created_at || '',
+    position: el.position,
+    type: 'Practice',
+    icon: 'strength',
+    task_type: 'EXERCISE',
+  })).sort((a, b) => b.position - a.position);
+
+  const updatedProject = project?.map((el) => {
+    const taskProps = getTaskProps(el?.slug?.slug || el?.slug);
+
+    return ({
+      ...el,
+      id,
+      label,
+      slug: el?.slug?.slug || el?.slug,
+      task_status: taskProps?.task_status || '',
+      revision_status: taskProps?.revision_status || '',
+      created_at: taskProps?.created_at || '',
+      daysDiff: taskProps?.created_at ? differenceInDays(currentDate, new Date(taskProps?.created_at)) : '',
+      position: el.position,
+      mandatory: el.mandatory,
+      type: 'Project',
+      icon: 'code',
+      task_type: 'PROJECT',
+    });
+  }).sort((a, b) => b.position - a.position);
+
+  const updatedAnswer = answer?.map((el) => ({
+    ...el,
+    id,
+    label,
+    task_status: getTaskProps(el.slug)?.task_status || '',
+    revision_status: getTaskProps(el.slug)?.revision_status || '',
+    created_at: getTaskProps(el.slug)?.created_at || '',
+    position: el.position,
+    type: 'Answer',
+    icon: 'answer',
+    task_type: 'QUIZ',
+  })).sort((a, b) => b.position - a.position);
+
+  const modules = [...updatedRead, ...updatedPractice, ...updatedProject, ...updatedAnswer];
+
+  const includesDailyTask = (module) => {
+    const getModules = taskTodo.some((task) => task.associated_slug === module.slug);
+    return getModules;
+  };
+
+  const includesStatusPending = (module) => {
+    const getModules = module.task_status === 'PENDING' && module.revision_status !== 'APPROVED';
+    return getModules;
+  };
+
+  const filteredModules = modules.filter((module) => includesDailyTask(module));
+  const filteredModulesByPending = modules.filter((module) => includesStatusPending(module));
+
+  return {
+    filteredModules,
+    modules,
+    filteredModulesByPending,
+  };
+  /*
+    example:
+    filteredModules: [{...}, {...}, {...}]
+    filteredModules: [{...}, {...}]
+    filteredModules: []
+    filteredModules: []
+    filteredModules: [{...}]
+    --------------------------------------------------
+    modules: [{...}, {...}, {...}]
+    modules: [{...}, {...}]
+    modules: [{...}, {...}, {...}, {...}, {...}]
+  */
+};
+
+export const getTechonologies = (cohortDays) => {
+  let technologyTags = [];
+
+  for (let i = 0; i < cohortDays.length; i += 1) {
+    if (typeof cohortDays[i].technologies === 'string') technologyTags.push(cohortDays[i].technologies);
+    if (Array.isArray(cohortDays[i].technologies)) {
+      technologyTags = technologyTags.concat(cohortDays[i].technologies);
+    }
+  }
+  technologyTags = [...new Set(technologyTags)];
+
+  return technologyTags;
+};
